+++ NEWS.rst
@@ -1,12 +1,12 @@
-News for storage
-================
+News for hex
+============
 
 This file lists the major changes between versions. For a more detailed list of
 every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Removed dependency on storage to ease reuseability.
 
 1.0.0
 -----